backend/ml/requirement.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
supabase
python-dotenv
